pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Игорь Рудич <rudi4.clarinet@mail.ru>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
dj-database-url = "^2.1.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
django-debug-toolbar = "^4.3.0"
django-widget-tweaks = "^1.5.0"
django-bootstrap5 = "^23.4"
django-filter = "^24.2"
rollbar = "^1.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
